FaaS_GCP/Batch_with_scheduler/delete_stories_expired/.requirements.txt
functions-framework==3.*
motor
python-dotenv